# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DataGuard",
    "summary": "Qualité et observabilité des données publiques",
    "description": "Ingère une source publique, exécute des contrôles de qualité, calcule un score expliqué et compare chaque exécution à la précédente.\n\nLe score est une aide à la lecture : il n'est jamais publié sans le détail des contrôles qui le composent, et les poids employés sont recopiés dans chaque résultat.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Lancer une exécution depuis l'URL de la source",
        "description": "Exécute le pipeline de bout en bout.\n\nLe traitement est synchrone : un dataset de quelques mégaoctets se traite en\nquelques secondes, ce qui ne justifie pas d'introduire une file de tâches.",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/upload": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Lancer une exécution à partir d'un fichier envoyé",
        "operationId": "create_run_from_upload_api_v1_runs_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_from_upload_api_v1_runs_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/results": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Results",
        "operationId": "get_results_api_v1_runs__run_id__results_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityResultOut"
                  },
                  "title": "Response Get Results Api V1 Runs  Run Id  Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/schema": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Schema",
        "operationId": "get_schema_api_v1_runs__run_id__schema_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchemaFieldOut"
                  },
                  "title": "Response Get Schema Api V1 Runs  Run Id  Schema Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/comparison": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Comparison",
        "operationId": "get_comparison_api_v1_runs__run_id__comparison_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunComparisonOut"
                }
              }
            }
          },
          "204": {
            "description": "Première exécution : rien à comparer"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/report": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Rapport JSON complet, téléchargeable",
        "operationId": "get_report_api_v1_runs__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_v1_datasets__dataset_id__runs_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RunSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_run_from_upload_api_v1_runs_upload_post": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "file"
        ],
        "title": "Body_create_run_from_upload_api_v1_runs_upload_post"
      },
      "CheckDiff": {
        "properties": {
          "check_name": {
            "type": "string",
            "title": "Check Name"
          },
          "from": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "success",
                  "warning",
                  "failed",
                  "skipped"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          },
          "to": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "success",
                  "warning",
                  "failed",
                  "skipped"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "check_name",
          "from",
          "to"
        ],
        "title": "CheckDiff"
      },
...
      "Freshness": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "age_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Hours"
          }
        },
        "type": "object",
        "required": [
          "date",
          "origin",
          "age_hours"
        ],
        "title": "Freshness"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_RunSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RunSummary]"
      },
      "QualityResultOut": {
        "properties": {
          "check_name": {
            "type": "string",
            "title": "Check Name"
          },
          "dimension": {
            "type": "string",
            "title": "Dimension"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "warning",
              "failed",
              "skipped"
            ],
            "title": "Status"
          },
          "weight": {
            "type": "integer",
            "title": "Weight"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "threshold": {
            "additionalProperties": true,
            "type": "object",
            "title": "Threshold"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "check_name",
          "dimension",
          "status",
          "weight",
          "score",
          "failed_count",
          "threshold",
          "details"
        ],
        "title": "QualityResultOut"
      },
      "RenamedColumn": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From"
          },
          "to": {
            "type": "string",
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to"
        ],
        "title": "RenamedColumn"
      },
      "RunComparisonOut": {
        "properties": {
          "previous_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Previous Run Id"
          },
          "rows_delta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Delta"
          },
          "rows_delta_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Delta Pct"
          },
          "score_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Delta"
          },
          "schema_diff": {
            "$ref": "#/components/schemas/SchemaDiff"
          },
          "checks_diff": {
            "items": {
              "$ref": "#/components/schemas/CheckDiff"
            },
            "type": "array",
            "title": "Checks Diff"
          }
        },
        "type": "object",
        "required": [
          "previous_run_id",
          "rows_delta",
          "rows_delta_pct",
          "score_delta",
          "schema_diff",
          "checks_diff"
        ],
        "title": "RunComparisonOut"
      },
      "RunCreate": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "RunCreate"
      },
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "warning",
              "failed",
              "skipped"
            ],
            "title": "Status"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "rows_read": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Read"
          },
          "columns_read": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns Read"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/QualityResultOut"
            },
            "type": "array",
            "title": "Results"
          },
          "schema": {
            "items": {
              "$ref": "#/components/schemas/SchemaFieldOut"
            },
            "type": "array",
            "title": "Schema"
          },
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunComparisonOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "started_at",
          "finished_at",
          "duration_ms",
          "status",
          "score",
          "rows_read",
          "columns_read",
          "error_message",
          "file_size_bytes",
          "checksum",
          "freshness",
          "results",
          "schema",
          "comparison"
        ],
        "title": "RunDetail"
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "warning",
              "failed",
              "skipped"
            ],
            "title": "Status"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "rows_read": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Read"
          },
          "columns_read": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns Read"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "started_at",
          "finished_at",
          "duration_ms",
          "status",
          "score",
          "rows_read",
          "columns_read",
          "error_message"
        ],
        "title": "RunSummary"
      },
      "SchemaDiff": {
        "properties": {
          "added": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Added"
          },
          "removed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed"
          },
          "renamed": {
            "items": {
              "$ref": "#/components/schemas/RenamedColumn"
            },
            "type": "array",
            "title": "Renamed"
          },
          "type_changed": {
            "items": {
              "$ref": "#/components/schemas/TypeChangedColumn"
            },
            "type": "array",
            "title": "Type Changed"
          }
        },
        "type": "object",
        "required": [
          "added",
          "removed",
          "renamed",
          "type_changed"
        ],
        "title": "SchemaDiff",
        "description": "Différences de schéma entre deux exécutions.\n\nDécrit explicitement plutôt qu'en objet libre : c'est ce qui permet au\ndashboard d'être vérifié par le compilateur au lieu de deviner la forme."
      },
      "SchemaFieldOut": {
        "properties": {
          "column_name": {
            "type": "string",
            "title": "Column Name"
          },
          "data_type": {
            "type": "string",
            "title": "Data Type"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "column_name",
          "data_type",
          "nullable",
          "position"
        ],
        "title": "SchemaFieldOut"
      },
...
      "TypeChangedColumn": {
        "properties": {
          "column": {
            "type": "string",
            "title": "Column"
          },
          "from": {
            "type": "string",
            "title": "From"
          },
          "to": {
            "type": "string",
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "column",
          "from",
          "to"
        ],
        "title": "TypeChangedColumn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}